refactor(memory): split ironclaw_memory into agnostic layer + native provider
Carve ironclaw_memory into two crates so memory is modeled as a pluggable
provider behind a shared interface (issue #3537):

- ironclaw_memory: the provider-agnostic memory layer — the MemoryService
  interface + operation DTOs, scope/path/context value types, prompt-write
  safety vocabulary, and audit/event contracts. What the host and every
  provider share.
- ironclaw_memory_native (new): the native filesystem provider —
  NativeMemoryService, backend/repository, the /memory adapter, chunking/
  search/indexing, and the prompt-write-safety enforcement engine. One
  provider, a peer to a future ironclaw_memory_honcho. Depends on
  ironclaw_memory.

Consumers depend on ironclaw_memory for the interface; host_runtime and the
native crate's own tests use ironclaw_memory_native for the concrete provider.

Behavior-preserving — pure reorganization. ironclaw_memory and
ironclaw_memory_native test suites are green; host_runtime results are
identical to the M1 baseline (244 passed, same 5 pre-existing Docker-socket /
prompt_doc_ref fixture failures); the lone event_projections failure
(extension_lifecycle) is pre-existing and imports no memory code. The only
internal change is MemorySignificantEvent::search_performed, which now takes
full_text/vector primitives instead of &MemorySearchRequest so the agnostic
crate stays free of search-impl dependencies; it builds the identical event.
