+++ crates/ironclaw_memory/src/context.rs
@@ -1,0 +1,85 @@
+//! Host-resolved scoped memory context.
+
+use ironclaw_host_api::{CorrelationId, ResourceScope};
+
+use crate::events::MemoryAuditContext;
+use crate::path::MemoryDocumentScope;
+use crate::safety::PromptSafetyAllowanceId;
+
+/// Host-resolved scoped context passed to memory backends.
+///
+/// Backends receive this context after the host has parsed and authorized the
+/// virtual path. They must not infer broader tenant/user/project authority from
+/// their own configuration.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct MemoryContext {
+    scope: MemoryDocumentScope,
+    invocation_id: Option<String>,
+    audit_context: Option<MemoryAuditContext>,
+    prompt_write_safety_allowance: Option<PromptSafetyAllowanceId>,
+    prompt_write_safety_enforced: bool,
+}
+
+impl MemoryContext {
+    pub fn new(scope: MemoryDocumentScope) -> Self {
+        Self {
+            scope,
+            invocation_id: None,
+            audit_context: None,
+            prompt_write_safety_allowance: None,
+            prompt_write_safety_enforced: false,
+        }
+    }
+
+    pub fn with_invocation_id(mut self, invocation_id: impl Into<String>) -> Self {
+        self.invocation_id = Some(invocation_id.into());
+        self
+    }
+
+    pub fn with_audit_context(
+        mut self,
+        resource_scope: ResourceScope,
+        correlation_id: CorrelationId,
+    ) -> Self {
+        self.invocation_id = Some(resource_scope.invocation_id.to_string());
+        self.audit_context = Some(MemoryAuditContext::new(resource_scope, correlation_id));
+        self
+    }
+
+    pub fn with_prompt_write_safety_allowance(
+        mut self,
+        allowance: PromptSafetyAllowanceId,
+    ) -> Self {
+        self.prompt_write_safety_allowance = Some(allowance);
+        self
+    }
+
+    /// Internal marker set only after `MemoryBackendFilesystemAdapter` has
+    /// already run prompt-write safety. Direct backend callers must not set
+    /// this to bypass protected prompt-file policy for files called out
+    /// in zmanian #3180 HIGH, including `SOUL.md` and `BOOTSTRAP.md`.
+    pub fn with_prompt_write_safety_enforced(mut self) -> Self {
+        self.prompt_write_safety_enforced = true;
+        self
+    }
+
+    pub fn scope(&self) -> &MemoryDocumentScope {
+        &self.scope
+    }
+
+    pub fn invocation_id(&self) -> Option<&str> {
+        self.invocation_id.as_deref()
+    }
+
+    pub fn audit_context(&self) -> Option<&MemoryAuditContext> {
+        self.audit_context.as_ref()
+    }
+
+    pub fn prompt_write_safety_allowance(&self) -> Option<&PromptSafetyAllowanceId> {
+        self.prompt_write_safety_allowance.as_ref()
+    }
+
+    pub fn prompt_write_safety_enforced(&self) -> bool {
+        self.prompt_write_safety_enforced
+    }
+}
