+++ crates/ironclaw_memory/src/safety.rs
@@ -1,26 +1,20 @@
-//! Prompt-write safety policy primitives for the memory crate.
+//! Prompt-write safety contract vocabulary for the memory crate.
 //!
 //! Protected memory documents (system prompt, identity, profile, hygiene
 //! configuration) need a uniform safety boundary regardless of which write
 //! surface — adapter, filesystem, indexer — performs the mutation. This
-//! module owns the vocabulary (operation, source, severity, reason codes,
-//! event sink) and the enforcement helpers that the memory backend and the
-//! filesystem adapters call.
+//! module owns the provider-neutral vocabulary (operation, source, severity,
+//! reason codes, event sink, policy trait) that the memory backend and the
+//! filesystem adapters depend on. The default policy implementation and the
+//! enforcement engine live in the `ironclaw_memory` implementation crate.
 
 use std::collections::BTreeSet;
-use std::sync::{Arc, Mutex};
 
 use async_trait::async_trait;
-use ironclaw_filesystem::{FilesystemError, FilesystemOperation};
 use ironclaw_host_api::{HostApiError, VirtualPath};
-use ironclaw_safety::{Sanitizer, Severity};
 
-use crate::chunking::content_sha256;
 use crate::events::{MemoryAuditContext, MemoryEventSinkError};
-use crate::path::{
-    MemoryDocumentPath, MemoryDocumentScope, memory_error, valid_memory_path,
-    validated_memory_relative_path,
-};
+use crate::path::{MemoryDocumentPath, MemoryDocumentScope, validated_memory_relative_path};
 
 /// Version identifier for the protected prompt-path policy registry.
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -279,17 +273,6 @@ impl PromptSafetySeverity {
     }
 }
 
-impl From<Severity> for PromptSafetySeverity {
-    fn from(severity: Severity) -> Self {
-        match severity {
-            Severity::Low => Self::Low,
-            Severity::Medium => Self::Medium,
-            Severity::High => Self::High,
-            Severity::Critical => Self::Critical,
-        }
-    }
-}
-
 /// Sanitized finding summary. It never includes raw content, matched text, or detector descriptions.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct PromptSafetySummary {
@@ -339,7 +322,7 @@ pub struct PromptSafetyReason {
 }
 
 impl PromptSafetyReason {
-    fn new(code: PromptSafetyReasonCode) -> Self {
+    pub fn new(code: PromptSafetyReasonCode) -> Self {
         Self {
             code,
             severity: None,
@@ -348,7 +331,7 @@ impl PromptSafetyReason {
         }
     }
 
-    fn with_findings(
+    pub fn with_findings(
         code: PromptSafetyReasonCode,
         severity: PromptSafetySeverity,
         finding_count: usize,
@@ -451,413 +434,6 @@ pub trait PromptWriteSafetyPolicy: Send + Sync {
         &self,
         request: PromptWriteSafetyRequest<'_>,
     ) -> Result<PromptWriteSafetyDecision, PromptWriteSafetyError>;
-}
-
-/// Default prompt-write safety policy preserving current workspace scanner behavior.
-pub struct DefaultPromptWriteSafetyPolicy {
-    registry: PromptProtectedPathRegistry,
-    sanitizer: Sanitizer,
-}
-
-impl DefaultPromptWriteSafetyPolicy {
-    pub fn new() -> Self {
-        Self::with_registry(PromptProtectedPathRegistry::default())
-    }
-
-    pub fn with_registry(registry: PromptProtectedPathRegistry) -> Self {
-        Self {
-            registry,
-            sanitizer: Sanitizer::new(),
-        }
-    }
-}
-
-impl Default for DefaultPromptWriteSafetyPolicy {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[async_trait]
-impl PromptWriteSafetyPolicy for DefaultPromptWriteSafetyPolicy {
-    fn protected_path_registry(&self) -> Option<&PromptProtectedPathRegistry> {
-        Some(&self.registry)
-    }
-
-    async fn check_write(
-        &self,
-        request: PromptWriteSafetyRequest<'_>,
-    ) -> Result<PromptWriteSafetyDecision, PromptWriteSafetyError> {
-        let protected_path_class = request.protected_path_class.cloned().or_else(|| {
-            request
-                .relative_memory_path
-                .and_then(|path| self.registry.classify_relative_path(path))
-        });
-        let Some(protected_path_class) = protected_path_class else {
-            return Ok(PromptWriteSafetyDecision::Allow);
-        };
-
-        if request.content.trim().is_empty() {
-            if let Some(allowance) = request.allowance
-                && *allowance == PromptSafetyAllowanceId::empty_prompt_file_clear()
-            {
-                return Ok(PromptWriteSafetyDecision::BypassAllowed {
-                    allowance: allowance.clone(),
-                });
-            }
-            return Ok(PromptWriteSafetyDecision::Reject {
-                reason: PromptSafetyReason {
-                    protected_path_class: Some(protected_path_class),
-                    ..PromptSafetyReason::new(PromptSafetyReasonCode::PromptWriteBypassNotAllowed)
-                },
-            });
-        }
-
-        let warnings = self.sanitizer.detect(request.content);
-        let Some(max_severity) = warnings.iter().map(|warning| warning.severity).max() else {
-            return Ok(PromptWriteSafetyDecision::Allow);
-        };
-        let severity = PromptSafetySeverity::from(max_severity);
-        let finding_count = warnings.len();
-
-        if max_severity >= Severity::Critical {
-            return Ok(PromptWriteSafetyDecision::Reject {
-                reason: PromptSafetyReason::with_findings(
-                    PromptSafetyReasonCode::CriticalPromptInjection,
-                    severity,
-                    finding_count,
-                    Some(protected_path_class),
-                ),
-            });
-        }
-        if max_severity >= Severity::High {
-            return Ok(PromptWriteSafetyDecision::Reject {
-                reason: PromptSafetyReason::with_findings(
-                    PromptSafetyReasonCode::HighRiskPromptInjection,
-                    severity,
-                    finding_count,
-                    Some(protected_path_class),
-                ),
-            });
-        }
-
-        Ok(PromptWriteSafetyDecision::Warn {
-            findings: PromptSafetySummary {
-                severity,
-                finding_count,
-            },
-        })
-    }
-}
-
-pub(crate) fn prompt_write_protected_classification(
-    policy: Option<&Arc<dyn PromptWriteSafetyPolicy>>,
-    registry: &PromptProtectedPathRegistry,
-    path: &MemoryDocumentPath,
-) -> Option<(PromptProtectedPathClass, PromptSafetyPolicyVersion)> {
-    if let Some(path_class) = registry.classify_path(path) {
-        return Some((path_class, registry.policy_version().clone()));
-    }
-    policy
-        .and_then(|policy| policy.protected_path_registry())
-        .and_then(|registry| {
-            registry
-                .classify_path(path)
-                .map(|path_class| (path_class, registry.policy_version().clone()))
-        })
-}
-
-pub(crate) fn prompt_write_policy_requires_previous_content_hash(
-    policy: Option<&Arc<dyn PromptWriteSafetyPolicy>>,
-) -> bool {
-    policy
-        .map(|policy| policy.requires_previous_content_hash())
-        .unwrap_or(false)
-}
-
-pub(crate) struct PromptWriteSafetyCheck<'a> {
-    pub scope: &'a MemoryDocumentScope,
-    pub path: &'a MemoryDocumentPath,
-    pub operation: PromptWriteOperation,
-    pub source: PromptWriteSource,
-    pub content: &'a str,
-    pub previous_content_hash: Option<&'a str>,
-    pub allowance: Option<&'a PromptSafetyAllowanceId>,
-    pub audit_context: Option<&'a MemoryAuditContext>,
-    pub filesystem_operation: FilesystemOperation,
-}
-
-#[derive(Debug, Clone, Default)]
-pub(crate) struct PromptWriteSafetyEnforcement {
-    pub allowance: Option<PromptSafetyAllowanceId>,
-}
-
-pub(crate) async fn enforce_prompt_write_safety(
-    policy: Option<&Arc<dyn PromptWriteSafetyPolicy>>,
-    event_sink: Option<&Arc<dyn PromptWriteSafetyEventSink>>,
-    registry: &PromptProtectedPathRegistry,
-    check: PromptWriteSafetyCheck<'_>,
-) -> Result<PromptWriteSafetyEnforcement, FilesystemError> {
-    let Some((protected_path_class, policy_version)) =
-        prompt_write_protected_classification(policy, registry, check.path)
-    else {
-        return Ok(PromptWriteSafetyEnforcement::default());
-    };
-    let virtual_path = check
-        .path
-        .virtual_path()
-        .unwrap_or_else(|_| valid_memory_path());
-    let Some(policy) = policy else {
-        let reason = PromptSafetyReason::new(PromptSafetyReasonCode::PromptWritePolicyUnavailable);
-        emit_prompt_write_safety_event(
-            event_sink,
-            &check,
-            PromptWriteSafetyEventParts {
-                kind: PromptWriteSafetyEventKind::Rejected,
-                policy_version: &policy_version,
-                protected_path_class: &protected_path_class,
-                reason: Some(&reason),
-                findings: None,
-                allowance: None,
-                require_sink: false,
-            },
-        )
-        .await?;
-        return Err(prompt_write_safety_error(
-            virtual_path,
-            check.filesystem_operation,
-            reason,
-        ));
-    };
-
-    let request = PromptWriteSafetyRequest {
-        scope: check.scope,
-        path: &virtual_path,
-        relative_memory_path: Some(check.path.relative_path()),
-        operation: check.operation,
-        source: check.source,
-        content: check.content,
-        previous_content_hash: check.previous_content_hash,
-        policy_version: policy_version.clone(),
-        protected_path_class: Some(&protected_path_class),
-        allowance: check.allowance,
-    };
-
-    match policy.check_write(request).await {
-        Ok(PromptWriteSafetyDecision::Allow) => {
-            emit_prompt_write_safety_event(
-                event_sink,
-                &check,
-                PromptWriteSafetyEventParts {
-                    kind: PromptWriteSafetyEventKind::Checked,
-                    policy_version: &policy_version,
-                    protected_path_class: &protected_path_class,
-                    reason: None,
-                    findings: None,
-                    allowance: None,
-                    require_sink: false,
-                },
-            )
-            .await?;
-            Ok(PromptWriteSafetyEnforcement::default())
-        }
-        Ok(PromptWriteSafetyDecision::BypassAllowed { allowance }) => {
-            emit_prompt_write_safety_event(
-                event_sink,
-                &check,
-                PromptWriteSafetyEventParts {
-                    kind: PromptWriteSafetyEventKind::BypassAllowed,
-                    policy_version: &policy_version,
-                    protected_path_class: &protected_path_class,
-                    reason: None,
-                    findings: None,
-                    allowance: Some(&allowance),
-                    require_sink: true,
-                },
-            )
-            .await?;
-            tracing::debug!(
-                target: "ironclaw::memory::prompt_write_safety",
-                operation = %check.operation,
-                source = %check.source,
-                protected_path_class = %protected_path_class.as_str(),
-                policy_version = %policy_version,
-                allowance = %allowance,
-                "protected prompt write bypass allowed"
-            );
-            Ok(PromptWriteSafetyEnforcement {
-                allowance: Some(allowance),
-            })
-        }
-        Ok(PromptWriteSafetyDecision::Warn { findings }) => {
-            emit_prompt_write_safety_event(
-                event_sink,
-                &check,
-                PromptWriteSafetyEventParts {
-                    kind: PromptWriteSafetyEventKind::Warned,
-                    policy_version: &policy_version,
-                    protected_path_class: &protected_path_class,
-                    reason: None,
-                    findings: Some(&findings),
-                    allowance: None,
-                    require_sink: true,
-                },
-            )
-            .await?;
-            tracing::debug!(
-                target: "ironclaw::memory::prompt_write_safety",
-                operation = %check.operation,
-                source = %check.source,
-                protected_path_class = %protected_path_class.as_str(),
-                policy_version = %policy_version,
-                severity = %findings.severity.as_str(),
-                finding_count = findings.finding_count,
-                "protected prompt write allowed with sanitized safety warning"
-            );
-            Ok(PromptWriteSafetyEnforcement::default())
-        }
-        Ok(PromptWriteSafetyDecision::Reject { reason }) => {
-            emit_prompt_write_safety_event(
-                event_sink,
-                &check,
-                PromptWriteSafetyEventParts {
-                    kind: PromptWriteSafetyEventKind::Rejected,
-                    policy_version: &policy_version,
-                    protected_path_class: &protected_path_class,
-                    reason: Some(&reason),
-                    findings: None,
-                    allowance: None,
-                    require_sink: false,
-                },
-            )
-            .await?;
-            Err(prompt_write_safety_error(
-                virtual_path,
-                check.filesystem_operation,
-                reason,
-            ))
-        }
-        Err(error) => {
-            let reason = error.reason;
-            emit_prompt_write_safety_event(
-                event_sink,
-                &check,
-                PromptWriteSafetyEventParts {
-                    kind: PromptWriteSafetyEventKind::Rejected,
-                    policy_version: &policy_version,
-                    protected_path_class: &protected_path_class,
-                    reason: Some(&reason),
-                    findings: None,
-                    allowance: None,
-                    require_sink: false,
-                },
-            )
-            .await?;
-            Err(prompt_write_safety_error(
-                virtual_path,
-                check.filesystem_operation,
-                reason,
-            ))
-        }
-    }
-}
-
-struct PromptWriteSafetyEventParts<'a> {
-    kind: PromptWriteSafetyEventKind,
-    policy_version: &'a PromptSafetyPolicyVersion,
-    protected_path_class: &'a PromptProtectedPathClass,
-    reason: Option<&'a PromptSafetyReason>,
-    findings: Option<&'a PromptSafetySummary>,
-    allowance: Option<&'a PromptSafetyAllowanceId>,
-    // Outcomes that would still persist with a non-clean safety result (warn/bypass)
-    // require a durable redacted audit seam before persistence.
-    require_sink: bool,
-}
-
-async fn emit_prompt_write_safety_event(
-    event_sink: Option<&Arc<dyn PromptWriteSafetyEventSink>>,
-    check: &PromptWriteSafetyCheck<'_>,
-    parts: PromptWriteSafetyEventParts<'_>,
-) -> Result<(), FilesystemError> {
-    let Some(event_sink) = event_sink else {
-        return if parts.require_sink {
-            Err(prompt_write_safety_error(
-                check
-                    .path
-                    .virtual_path()
-                    .unwrap_or_else(|_| valid_memory_path()),
-                check.filesystem_operation,
-                PromptSafetyReason::new(PromptSafetyReasonCode::PromptWriteSafetyEventUnavailable),
-            ))
-        } else {
-            Ok(())
-        };
-    };
-    let event = PromptWriteSafetyEvent {
-        kind: parts.kind,
-        scope: check.scope.clone(),
-        operation: check.operation,
-        source: check.source,
-        policy_version: parts.policy_version.clone(),
-        protected_path_class: Some(parts.protected_path_class.clone()),
-        relative_path_hash: Some(content_sha256(check.path.relative_path())),
-        reason_code: parts.reason.map(|reason| reason.code),
-        severity: parts
-            .reason
-            .and_then(|reason| reason.severity)
-            .or_else(|| parts.findings.map(|findings| findings.severity)),
-        finding_count: parts
-            .reason
-            .map(|reason| reason.finding_count)
-            .or_else(|| parts.findings.map(|findings| findings.finding_count))
-            .unwrap_or(0),
-        allowance: parts.allowance.cloned(),
-        audit_context: check.audit_context.cloned(),
-    };
-    if let Err(error) = event_sink.record_prompt_write_safety_event(event).await {
-        tracing::debug!(
-            target: "ironclaw::memory::prompt_write_safety",
-            error = %error,
-            operation = %check.operation,
-            source = %check.source,
-            "failed to record prompt write safety event"
-        );
-        if parts.require_sink {
-            return Err(prompt_write_safety_error(
-                check
-                    .path
-                    .virtual_path()
-                    .unwrap_or_else(|_| valid_memory_path()),
-                check.filesystem_operation,
-                PromptSafetyReason::new(PromptSafetyReasonCode::PromptWriteSafetyEventUnavailable),
-            ));
-        }
-        return Ok(());
-    }
-    Ok(())
-}
-
-fn prompt_write_safety_error(
-    path: VirtualPath,
-    operation: FilesystemOperation,
-    reason: PromptSafetyReason,
-) -> FilesystemError {
-    memory_error(path, operation, reason.code.as_str())
-}
-
-pub(crate) fn take_prompt_safety_allowance(
-    allowance: &Mutex<Option<PromptSafetyAllowanceId>>,
-    path: &VirtualPath,
-    operation: FilesystemOperation,
-) -> Result<Option<PromptSafetyAllowanceId>, FilesystemError> {
-    let mut allowance = allowance.lock().map_err(|_| {
-        memory_error(
-            path.clone(),
-            operation,
-            "prompt write safety allowance lock poisoned",
-        )
-    })?;
-    Ok(allowance.take())
 }
 
 #[cfg(test)]
