+++ crates/ironclaw_memory_native/CLAUDE.md
