+++ crates/ironclaw_memory_native/src/contract_tests.rs
