+++ crates/ironclaw_memory_native/src/embedding.rs
