+++ crates/ironclaw_memory_native/src/indexer.rs
