+++ crates/ironclaw_memory_native/src/schema.rs
