+++ crates/ironclaw_memory_native/src/search.rs
