+++ crates/ironclaw_memory_native/src/service.rs
@@ -1,0 +1,726 @@
+//! IronClaw memory service facade for Reborn.
+//!
+//! This module owns the host-facing IronClaw memory operation shapes. Host
+//! runtime callers still resolve scope, mounts, grants, approvals, and audit
+//! services before calling the service; the default native adapter keeps the
+//! existing storage format.
+
+use std::{cmp::Ordering, collections::BTreeMap, sync::Arc};
+
+use crate::{
+    ChunkingMemoryDocumentIndexer, DocumentMetadata, FilesystemMemoryDocumentRepository,
+    MemoryBackend, MemoryBackendCapabilities, MemoryBackendWriteOptions, MemoryContext,
+    MemoryDocumentPath, MemoryDocumentScope, MemorySearchRequest, MemorySearchResult,
+    MemoryWriteOutcome, PromptSafetyAllowanceId, PromptWriteSafetyEventSink,
+    RepositoryMemoryBackend, content_bytes_sha256,
+};
+use async_trait::async_trait;
+use chrono::Utc;
+use chrono_tz::Tz;
+use ironclaw_filesystem::RootFilesystem;
+use ironclaw_prompt_envelope::{EnvelopeSource, EnvelopeTrust, wrap_untrusted_with_limit};
+use serde_json::{Map, Value, json};
+
+// The host-facing operation shapes + the `MemoryService` trait moved to
+// `ironclaw_memory`; re-exported so `crate::service::*` and the crate's
+// public API stay unchanged while `NativeMemoryService` (below) keeps the native
+// adapter behavior here.
+pub use ironclaw_memory::{
+    MemoryInvocation, MemoryService, MemoryServiceContextRequest, MemoryServiceContextSnippet,
+    MemoryServiceError, MemoryServiceErrorKind, MemoryServiceProfileSetRequest,
+    MemoryServiceProfileSetResponse, MemoryServiceReadRequest, MemoryServiceReadResponse,
+    MemoryServiceSearchRequest, MemoryServiceSearchResponse, MemoryServiceSearchResult,
+    MemoryServiceTreeRequest, MemoryServiceTreeResponse, MemoryServiceWriteRequest,
+    MemoryServiceWriteResponse,
+};
+
+const MEMORY_PATH: &str = "MEMORY.md";
+const HEARTBEAT_PATH: &str = "HEARTBEAT.md";
+const BOOTSTRAP_PATH: &str = "BOOTSTRAP.md";
+const PROFILE_DOCUMENT_PATH: &str = "context/profile.json";
+const MAX_MEMORY_PATCH_RETRIES: usize = 8;
+const MAX_SAFE_SUMMARY_BYTES: usize = 512;
+const MAX_TOTAL_SAFE_SUMMARY_BYTES: usize = 4 * 1024;
+const FNV_OFFSET: u64 = 0xcbf29ce484222325;
+const FNV_PRIME: u64 = 0x00000100000001B3;
+const FIELD_SEPARATOR: u8 = 0xFF;
+
+pub struct NativeMemoryService {
+    backend: Arc<dyn MemoryBackend>,
+}
+
+impl std::fmt::Debug for NativeMemoryService {
+    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        formatter
+            .debug_struct("NativeMemoryService")
+            .field("backend", &"<native-memory-backend>")
+            .finish()
+    }
+}
+
+impl NativeMemoryService {
+    pub fn new(backend: Arc<dyn MemoryBackend>) -> Self {
+        Self { backend }
+    }
+
+    pub fn from_filesystem(
+        filesystem: Arc<dyn RootFilesystem>,
+        prompt_write_safety_event_sink: Option<Arc<dyn PromptWriteSafetyEventSink>>,
+    ) -> Self {
+        Self {
+            backend: build_native_backend(filesystem, prompt_write_safety_event_sink),
+        }
+    }
+
+    fn scoped_context(
+        &self,
+        invocation: &MemoryInvocation,
+    ) -> Result<(MemoryDocumentScope, MemoryContext), MemoryServiceError> {
+        let scope = MemoryDocumentScope::new_with_agent(
+            invocation.scope.tenant_id.as_str(),
+            invocation.scope.user_id.as_str(),
+            invocation.scope.agent_id.as_ref().map(|id| id.as_str()),
+            invocation.scope.project_id.as_ref().map(|id| id.as_str()),
+        )
+        .map_err(|_| MemoryServiceError::input())?;
+        let context = MemoryContext::new(scope.clone())
+            .with_audit_context(invocation.scope.clone(), invocation.correlation_id);
+        Ok((scope, context))
+    }
+}
+
+#[async_trait]
+impl MemoryService for NativeMemoryService {
+    async fn search(
+        &self,
+        invocation: MemoryInvocation,
+        request: MemoryServiceSearchRequest,
+    ) -> Result<MemoryServiceSearchResponse, MemoryServiceError> {
+        let (_, context) = self.scoped_context(&invocation)?;
+        let search_request = MemorySearchRequest::new(&request.query)
+            .map_err(|_| MemoryServiceError::input())?
+            .with_limit(request.limit)
+            .with_pre_fusion_limit(request.limit.max(20))
+            .with_vector(false);
+        let results = self
+            .backend
+            .search(&context, search_request)
+            .await
+            .map_err(|_| MemoryServiceError::operation())?
+            .into_iter()
+            .map(|result| MemoryServiceSearchResult {
+                is_hybrid_match: result.is_hybrid(),
+                content: result.snippet,
+                score: result.score,
+                path: result.path.relative_path().to_string(),
+            })
+            .collect();
+        Ok(MemoryServiceSearchResponse {
+            query: request.query,
+            results,
+        })
+    }
+
+    async fn write(
+        &self,
+        invocation: MemoryInvocation,
+        request: MemoryServiceWriteRequest,
+    ) -> Result<MemoryServiceWriteResponse, MemoryServiceError> {
+        reject_local_or_traversal_path(&request.target)?;
+        let (scope, context) = self.scoped_context(&invocation)?;
+        let resolved_path = resolve_target_path(&request.target, request.timezone.as_deref())?;
+        let path = document_path(&scope, &resolved_path)?;
+        let options = write_options(request.metadata.as_ref());
+
+        if request.target == "bootstrap" {
+            if path.relative_path() != BOOTSTRAP_PATH || resolved_path != BOOTSTRAP_PATH {
+                return Err(MemoryServiceError::operation());
+            }
+            let context = context.clone().with_prompt_write_safety_allowance(
+                PromptSafetyAllowanceId::empty_prompt_file_clear(),
+            );
+            self.backend
+                .write_document_with_backend_options(&context, &path, b"", &options)
+                .await
+                .map_err(|_| MemoryServiceError::operation())?;
+            return Ok(MemoryServiceWriteResponse {
+                status: "cleared".to_string(),
+                path: resolved_path,
+                append: false,
+                content_length: 0,
+                replacements: None,
+                message: Some("BOOTSTRAP.md cleared.".to_string()),
+            });
+        }
+
+        if let Some(old_string) = request.old_string.as_deref() {
+            if old_string.is_empty() {
+                return Err(MemoryServiceError::input());
+            }
+            let new_string = request
+                .new_string
+                .as_deref()
+                .ok_or_else(MemoryServiceError::input)?;
+            return self
+                .patch_document(PatchDocumentRequest {
+                    context: &context,
+                    path: &path,
+                    resolved_path: &resolved_path,
+                    options: &options,
+                    old_string,
+                    new_string,
+                    replace_all: request.replace_all,
+                })
+                .await;
+        }
+
+        if request.content.trim().is_empty() {
+            return Err(MemoryServiceError::input());
+        }
+        if request.append {
+            self.backend
+                .append_document_with_backend_options(
+                    &context,
+                    &path,
+                    request.content.as_bytes(),
+                    &options,
+                )
+                .await
+                .map_err(|_| MemoryServiceError::operation())?;
+        } else {
+            self.backend
+                .write_document_with_backend_options(
+                    &context,
+                    &path,
+                    request.content.as_bytes(),
+                    &options,
+                )
+                .await
+                .map_err(|_| MemoryServiceError::operation())?;
+        }
+
+        Ok(MemoryServiceWriteResponse {
+            status: "written".to_string(),
+            path: resolved_path,
+            append: request.append,
+            content_length: request.content.len(),
+            replacements: None,
+            message: None,
+        })
+    }
+
+    async fn read(
+        &self,
+        invocation: MemoryInvocation,
+        request: MemoryServiceReadRequest,
+    ) -> Result<MemoryServiceReadResponse, MemoryServiceError> {
+        reject_local_or_traversal_path(&request.path)?;
+        let (scope, context) = self.scoped_context(&invocation)?;
+        let path = document_path(&scope, &request.path)?;
+        let Some(bytes) = self
+            .backend
+            .read_document(&context, &path)
+            .await
+            .map_err(|_| MemoryServiceError::operation())?
+        else {
+            return Err(MemoryServiceError::input());
+        };
+        let content = String::from_utf8(bytes).map_err(|_| MemoryServiceError::operation())?;
+        Ok(MemoryServiceReadResponse {
+            path: path.relative_path().to_string(),
+            word_count: content.split_whitespace().count(),
+            content,
+        })
+    }
+
+    async fn tree(
+        &self,
+        invocation: MemoryInvocation,
+        request: MemoryServiceTreeRequest,
+    ) -> Result<MemoryServiceTreeResponse, MemoryServiceError> {
+        if !request.path.is_empty() {
+            reject_local_or_traversal_path(&request.path)?;
+        }
+        let (scope, context) = self.scoped_context(&invocation)?;
+        let mut paths = self
+            .backend
+            .list_documents(&context, &scope)
+            .await
+            .map_err(|_| MemoryServiceError::operation())?
+            .into_iter()
+            .map(|path| path.relative_path().to_string())
+            .collect::<Vec<_>>();
+        paths.sort();
+        Ok(MemoryServiceTreeResponse {
+            entries: tree_for_paths(&paths, request.path.trim_matches('/'), request.depth),
+        })
+    }
+
+    async fn profile_set(
+        &self,
+        invocation: MemoryInvocation,
+        request: MemoryServiceProfileSetRequest,
+    ) -> Result<MemoryServiceProfileSetResponse, MemoryServiceError> {
+        let (scope, path) = profile_scope_and_path(
+            invocation.scope.tenant_id.as_str(),
+            invocation.scope.user_id.as_str(),
+        )?;
+        let context = MemoryContext::new(scope)
+            .with_audit_context(invocation.scope.clone(), invocation.correlation_id);
+        let options = write_options(None);
+        for _ in 0..MAX_MEMORY_PATCH_RETRIES {
+            let current = self
+                .backend
+                .read_document(&context, &path)
+                .await
+                .map_err(|_| MemoryServiceError::operation())?;
+            let expected_hash = current.as_deref().map(content_bytes_sha256);
+            let mut doc: Map<String, Value> = match &current {
+                Some(bytes) => {
+                    serde_json::from_slice(bytes).map_err(|_| MemoryServiceError::operation())?
+                }
+                None => Map::new(),
+            };
+            for key in ["timezone", "locale", "location"] {
+                if let Some(value) = doc.get(key)
+                    && !value.is_string()
+                {
+                    return Err(MemoryServiceError::operation());
+                }
+            }
+            for (key, value) in &request.fields {
+                doc.insert(key.clone(), value.clone());
+            }
+            let bytes = serde_json::to_vec(&Value::Object(doc))
+                .map_err(|_| MemoryServiceError::operation())?;
+            let outcome = self
+                .backend
+                .compare_and_write_document_with_backend_options(
+                    &context,
+                    &path,
+                    expected_hash.as_deref(),
+                    &bytes,
+                    &options,
+                )
+                .await
+                .map_err(|_| MemoryServiceError::operation())?;
+            if outcome == MemoryWriteOutcome::Written {
+                return Ok(MemoryServiceProfileSetResponse {
+                    status: "ok".to_string(),
+                });
+            }
+        }
+        Err(MemoryServiceError::operation())
+    }
+
+    async fn retrieve_context(
+        &self,
+        invocation: MemoryInvocation,
+        request: MemoryServiceContextRequest,
+    ) -> Result<Vec<MemoryServiceContextSnippet>, MemoryServiceError> {
+        if request.max_snippets == 0 || memory_context_disabled(&request.context_profile_id) {
+            return Ok(Vec::new());
+        }
+        let (_, context) = self.scoped_context(&invocation)?;
+        let search_request = MemorySearchRequest::new(&request.query)
+            .map_err(|_| MemoryServiceError::input())?
+            .with_limit(request.max_snippets);
+        let mut results = self
+            .backend
+            .search(&context, search_request)
+            .await
+            .map_err(|_| MemoryServiceError::unavailable())?;
+        results.retain(|result| result.path.scope() == context.scope() && result.score.is_finite());
+        results.sort_by(compare_memory_search_results);
+
+        Ok(collect_context_snippets(
+            results,
+            request.max_snippets,
+            MAX_TOTAL_SAFE_SUMMARY_BYTES,
+        ))
+    }
+}
+
+impl NativeMemoryService {
+    async fn patch_document(
+        &self,
+        request: PatchDocumentRequest<'_>,
+    ) -> Result<MemoryServiceWriteResponse, MemoryServiceError> {
+        for _ in 0..MAX_MEMORY_PATCH_RETRIES {
+            let Some(bytes) = self
+                .backend
+                .read_document(request.context, request.path)
+                .await
+                .map_err(|_| MemoryServiceError::operation())?
+            else {
+                return Err(MemoryServiceError::operation());
+            };
+            let existing = String::from_utf8(bytes).map_err(|_| MemoryServiceError::operation())?;
+            let expected = content_bytes_sha256(existing.as_bytes());
+            let replacements = existing.matches(request.old_string).count();
+            if replacements == 0 {
+                return Err(MemoryServiceError::input());
+            }
+            let replacement_count = if request.replace_all { replacements } else { 1 };
+            let updated = if request.replace_all {
+                existing.replace(request.old_string, request.new_string)
+            } else {
+                existing.replacen(request.old_string, request.new_string, 1)
+            };
+            let outcome = self
+                .backend
+                .compare_and_write_document_with_backend_options(
+                    request.context,
+                    request.path,
+                    Some(&expected),
+                    updated.as_bytes(),
+                    request.options,
+                )
+                .await
+                .map_err(|_| MemoryServiceError::operation())?;
+            if outcome == MemoryWriteOutcome::Written {
+                return Ok(MemoryServiceWriteResponse {
+                    status: "patched".to_string(),
+                    path: request.resolved_path.to_string(),
+                    append: false,
+                    content_length: updated.len(),
+                    replacements: Some(replacement_count),
+                    message: None,
+                });
+            }
+        }
+        Err(MemoryServiceError::operation())
+    }
+}
+
+struct PatchDocumentRequest<'a> {
+    context: &'a MemoryContext,
+    path: &'a MemoryDocumentPath,
+    resolved_path: &'a str,
+    options: &'a MemoryBackendWriteOptions,
+    old_string: &'a str,
+    new_string: &'a str,
+    replace_all: bool,
+}
+
+fn build_native_backend(
+    filesystem: Arc<dyn RootFilesystem>,
+    prompt_write_safety_event_sink: Option<Arc<dyn PromptWriteSafetyEventSink>>,
+) -> Arc<dyn MemoryBackend> {
+    let repository = Arc::new(FilesystemMemoryDocumentRepository::new(filesystem));
+    let indexer = Arc::new(ChunkingMemoryDocumentIndexer::new(Arc::clone(&repository)));
+    let mut backend = RepositoryMemoryBackend::new(Arc::clone(&repository))
+        .with_indexer(indexer)
+        .with_capabilities(MemoryBackendCapabilities {
+            file_documents: true,
+            metadata: true,
+            versioning: true,
+            prompt_write_safety: true,
+            full_text_search: true,
+            delete: true,
+            transactions: true,
+            ..MemoryBackendCapabilities::default()
+        });
+    if let Some(prompt_write_safety_event_sink) = prompt_write_safety_event_sink {
+        backend = backend.with_prompt_write_safety_event_sink(prompt_write_safety_event_sink);
+    }
+    Arc::new(backend)
+}
+
+fn resolve_target_path(target: &str, timezone: Option<&str>) -> Result<String, MemoryServiceError> {
+    match target {
+        "memory" => Ok(MEMORY_PATH.to_string()),
+        "heartbeat" => Ok(HEARTBEAT_PATH.to_string()),
+        "bootstrap" => Ok(BOOTSTRAP_PATH.to_string()),
+        "daily_log" => {
+            let timezone = match timezone {
+                Some(value) => value
+                    .parse::<Tz>()
+                    .map_err(|_| MemoryServiceError::input())?,
+                None => Tz::UTC,
+            };
+            let now = Utc::now().with_timezone(&timezone);
+            Ok(format!("daily/{}.md", now.format("%Y-%m-%d")))
+        }
+        path => Ok(path.to_string()),
+    }
+}
+
+fn document_path(
+    scope: &MemoryDocumentScope,
+    relative_path: &str,
+) -> Result<MemoryDocumentPath, MemoryServiceError> {
+    MemoryDocumentPath::new_with_agent(
+        scope.tenant_id(),
+        scope.user_id(),
+        scope.agent_id(),
+        scope.project_id(),
+        relative_path,
+    )
+    .map_err(|_| MemoryServiceError::input())
+}
+
+fn profile_scope_and_path(
+    tenant_id: &str,
+    user_id: &str,
+) -> Result<(MemoryDocumentScope, MemoryDocumentPath), MemoryServiceError> {
+    let scope = MemoryDocumentScope::new_with_agent(tenant_id, user_id, None, None)
+        .map_err(|_| MemoryServiceError::input())?;
+    let path =
+        MemoryDocumentPath::new_with_agent(tenant_id, user_id, None, None, PROFILE_DOCUMENT_PATH)
+            .map_err(|_| MemoryServiceError::input())?;
+    Ok((scope, path))
+}
+
+fn write_options(metadata_overlay: Option<&DocumentMetadata>) -> MemoryBackendWriteOptions {
+    MemoryBackendWriteOptions {
+        metadata_overlay: metadata_overlay.cloned(),
+    }
+}
+
+fn reject_local_or_traversal_path(path: &str) -> Result<(), MemoryServiceError> {
+    if path.contains('\\') || looks_like_filesystem_path(path) || contains_traversal(path) {
+        return Err(MemoryServiceError::input());
+    }
+    Ok(())
+}
+
+fn contains_traversal(path: &str) -> bool {
+    path.split('/').any(|segment| segment == "..")
+}
+
+fn looks_like_filesystem_path(path: &str) -> bool {
+    if path.is_empty() {
+        return false;
+    }
+    if path.starts_with('/') || path.starts_with("~/") {
+        return true;
+    }
+    let bytes = path.as_bytes();
+    bytes.len() >= 3
+        && bytes[0].is_ascii_alphabetic()
+        && bytes[1] == b':'
+        && (bytes[2] == b'\\' || bytes[2] == b'/')
+}
+
+fn tree_for_paths(paths: &[String], root: &str, max_depth: usize) -> Vec<Value> {
+    let prefix = if root.is_empty() {
+        String::new()
+    } else {
+        format!("{}/", root.trim_matches('/'))
+    };
+    let mut children = BTreeMap::<String, Vec<String>>::new();
+    let mut files = Vec::new();
+    for path in paths {
+        let Some(remainder) = path.strip_prefix(&prefix) else {
+            continue;
+        };
+        if remainder.is_empty() {
+            continue;
+        }
+        if let Some((dir, _)) = remainder.split_once('/') {
+            children
+                .entry(dir.to_string())
+                .or_default()
+                .push(path.clone());
+        } else {
+            files.push(remainder.to_string());
+        }
+    }
+
+    let mut output = Vec::new();
+    for (dir, child_paths) in children {
+        let display = format!("{dir}/");
+        if max_depth <= 1 {
+            output.push(Value::String(display));
+        } else {
+            let child_root = if root.is_empty() {
+                dir
+            } else {
+                format!("{root}/{dir}")
+            };
+            let child_tree = tree_for_paths(&child_paths, &child_root, max_depth - 1);
+            if child_tree.is_empty() {
+                output.push(Value::String(display));
+            } else {
+                output.push(json!({ (display): child_tree }));
+            }
+        }
+    }
+    output.extend(files.into_iter().map(Value::String));
+    output
+}
+
+fn compare_memory_search_results(
+    left: &MemorySearchResult,
+    right: &MemorySearchResult,
+) -> Ordering {
+    right
+        .score
+        .total_cmp(&left.score)
+        .then_with(|| left.path.relative_path().cmp(right.path.relative_path()))
+}
+
+fn collect_context_snippets(
+    results: Vec<MemorySearchResult>,
+    max_snippets: usize,
+    max_total_bytes: usize,
+) -> Vec<MemoryServiceContextSnippet> {
+    let mut snippets = Vec::new();
+    let mut total_bytes = 0usize;
+
+    for result in results {
+        if snippets.len() >= max_snippets {
+            break;
+        }
+        let Some(snippet) = map_search_result_to_snippet(result) else {
+            continue;
+        };
+        let snippet_bytes = snippet.safe_summary.len();
+        if total_bytes.saturating_add(snippet_bytes) > max_total_bytes {
+            break;
+        }
+        total_bytes = total_bytes.saturating_add(snippet_bytes);
+        snippets.push(snippet);
+    }
+
+    snippets
+}
+
+fn map_search_result_to_snippet(result: MemorySearchResult) -> Option<MemoryServiceContextSnippet> {
+    let snippet_ref = memory_snippet_display_ref([
+        result.path.tenant_id(),
+        result.path.user_id(),
+        result.path.agent_id().unwrap_or(""),
+        result.path.project_id().unwrap_or(""),
+        result.path.relative_path(),
+    ]);
+    let model_content = sanitize_snippet_text(&result.snippet)?;
+    Some(MemoryServiceContextSnippet {
+        snippet_ref,
+        safe_summary: model_content.clone(),
+        model_content,
+    })
+}
+
+fn memory_snippet_display_ref<'a>(parts: impl IntoIterator<Item = &'a str>) -> String {
+    let mut hash = FNV_OFFSET;
+    for field in parts {
+        feed_hash(&mut hash, field.as_bytes());
+        feed_hash(&mut hash, &[FIELD_SEPARATOR]);
+    }
+    format!("memory-snippet:{hash:016x}")
+}
+
+fn feed_hash(hash: &mut u64, bytes: &[u8]) {
+    for &byte in bytes {
+        *hash ^= u64::from(byte);
+        *hash = hash.wrapping_mul(FNV_PRIME);
+    }
+}
+
+fn sanitize_snippet_text(raw: &str) -> Option<String> {
+    const PROBE_BODY: &str = "x";
+    let probe = wrap_untrusted_with_limit(
+        EnvelopeSource::Memory,
+        EnvelopeTrust::Untrusted,
+        PROBE_BODY,
+        MAX_SAFE_SUMMARY_BYTES,
+    )
+    .ok()?;
+    let prefix_len = probe.byte_len().saturating_sub(PROBE_BODY.len());
+
+    let cleaned: String = raw.chars().filter(|ch| !ch.is_control()).collect();
+    let cleaned = cleaned.trim();
+    if cleaned.is_empty() {
+        return None;
+    }
+
+    let max_payload_bytes = MAX_SAFE_SUMMARY_BYTES.saturating_sub(prefix_len);
+    let truncated = truncate_to_char_boundary(cleaned, max_payload_bytes);
+    if truncated.is_empty() {
+        return None;
+    }
+
+    let envelope = wrap_untrusted_with_limit(
+        EnvelopeSource::Memory,
+        EnvelopeTrust::Untrusted,
+        truncated,
+        MAX_SAFE_SUMMARY_BYTES,
+    )
+    .ok()?
+    .into_string();
+    validate_loop_safe_summary(envelope)
+}
+
+fn truncate_to_char_boundary(value: &str, max_bytes: usize) -> &str {
+    if value.len() <= max_bytes {
+        return value;
+    }
+
+    let mut end = max_bytes;
+    while end > 0 && !value.is_char_boundary(end) {
+        end -= 1;
+    }
+    &value[..end]
+}
+
+fn memory_context_disabled(context_profile_id: &str) -> bool {
+    const MEMORY_DISABLED_ALIASES: &[&str] = &[
+        "memory_disabled",
+        "memory-disabled",
+        "disabled_context",
+        "context_disabled",
+    ];
+    MEMORY_DISABLED_ALIASES.contains(&context_profile_id)
+}
+
+fn validate_loop_safe_summary(value: String) -> Option<String> {
+    if value.is_empty()
+        || value.len() > MAX_SAFE_SUMMARY_BYTES
+        || value
+            .chars()
+            .any(|character| character == '\0' || character.is_control())
+        || value.chars().any(|character| {
+            matches!(
+                character,
+                '{' | '}' | '[' | ']' | '`' | '<' | '>' | '/' | '\\'
+            )
+        })
+    {
+        return None;
+    }
+
+    let lower = value.to_ascii_lowercase();
+    for forbidden in [
+        "access token",
+        "api key",
+        "api_key",
+        "apikey",
+        "authorization:",
+        "bearer ",
+        "host path",
+        "invalid api key",
+        "invalid_api_key",
+        "password",
+        "passwd",
+        "provider error",
+        "raw runtime",
+        "secret",
+        "stack trace",
+        "tool input",
+        "tool_input",
+        "traceback",
+    ] {
+        if lower.contains(forbidden) {
+            return None;
+        }
+    }
+    if lower
+        .split(|character: char| !character.is_ascii_alphanumeric() && character != '-')
+        .any(|token| token.starts_with("sk-"))
+    {
+        return None;
+    }
+    Some(value)
+}
