+++ crates/ironclaw_memory_native/src/write_metadata.rs
