+++ crates/ironclaw_memory_native/tests/memory_backend_contract.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 use async_trait::async_trait;
 use ironclaw_filesystem::{FilesystemError, FilesystemOperation, RootFilesystem};
 use ironclaw_host_api::VirtualPath;
-use ironclaw_memory::{
+use ironclaw_memory_native::{
     ChunkConfig, DefaultPromptWriteSafetyPolicy, DocumentMetadata,
     InMemoryMemoryDocumentRepository, MemoryAppendOutcome, MemoryBackend,
     MemoryBackendCapabilities, MemoryBackendFilesystemAdapter, MemoryBackendWriteOptions,
